fix review comment
